Integrate Cookie Consent plugin for cookie pop-up
Integrate Cookie Consent (https://silktide.com/tools/cookie-consent/)
with the theme. This is a plugin for displaying a banner/pop-up
informing a user that cookies are being used, to support compliance
with data protection rules.

Update README.md to show the configuration settings to control
the behaviour of the plugin.